+++ nomad-dash/src/components/brand.js
@@ -1,0 +1,10 @@
+import React from 'react'
+
+// Logo-brand for the Nomad interface
+const Brand = () => (
+  <span id="brand" className="text-uppercase mx-3">
+    NOMAD
+  </span>
+)
+
+export default Brand