+++ nomad-rpi/start.sh
@@ -1,0 +1,2 @@
+# Start the nomad-rpi program
+python3 ./brain.py