+++ CHANGELOG.md
@@ -7,7 +7,15 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased] - DD-MAY-2018
 ### Changed
 - `nomadRPi` -> `nomad-rpi`
+- `nomadFrontend` -> `nomad-dash`
 
+### Removed
+- `nomadVR`
+- `nomadBackend`
+- `nomadBrain`
+
+
+&nbsp;
 ## 0.0.1 - 17-MAR-2018
 ### Features
 - Firmware cartesian axes control
@@ -16,3 +24,5 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - UI interface developed using react with mock data
 - RPi sub-control functions in python
 - Modified firmware for increased speeds
+
+[Unreleased]: https://github.com/attackle/nomad/comapare